[duts] Add explicit date option
Proper date is important while testing something on target, in order
to have more reliable logs. This option makes it possible to pass
required date to duts as a param, so that it could be used by
testcases to reset target's date to that value. This will help when
"common recipe" - ntpd is not the best solution.

Since only ltp is going to make use of it, it will just set global
variable $curdate. In the future, if found appropriate, date sync
could be integrated into VL_ calls.

Signed-off-by: Vitaly Bordug <vitb@kernel.crashing.org>