duts: Add '-continue' option skipping already PASSed test cases.
As the summary report now only gets results appended, rather than
rewriting it every time, we can use it as a state for a test session.

The new option uses a function to parse the summary report and filters
the already passed tests from the list of test cases to run.

Signed-off-by: Detlev Zundel <dzu@denx.de>