context-handlers: Remove unneccessary prompts in logs.
The context handlers need not check for a new prompt before executing
a single command.  Instead we require a prompt once after logging has
been enabled.  Because every command only completes when seeing a
prompt, no check is needed before the next command.

This compacts the logfiles by removin unneccessary "empty prompts".

Still an empty line in the logs cannot be avoided, so we filter that
in dulg_verbatimize.

Signed-off-by: Detlev Zundel <dzu@denx.de>