testsystems/dulg/testcases: First stab at reordering testcases.
The testcases need to be organized from a perspective of testing a
system, not from generating documentation for it.  The latter will
simply be possible if all aspects have been tested.

This reorganization is in now way considered to be complete, but it
should show the direction we want to take.

Signed-off-by: Detlev Zundel <dzu@denx.de>