testsystems/dulg: Fix LinuxMtd to work with NANDs with notable pagesize
The testcase used to expose chardev interface of MTD device with cat/echo
small chunks of data there. It will not work with modern NANDs, with
special parameter - minimum I/O unit size, or NAND page size.

Unfortunately there is no clear way to figure that out from sysfs/whatever,
and so far mtd FAQs referring to  the flash manual (it is usually 512, 2048
or even 4096 at some modern NANDs). It is apparently 1 for NOR flash, unless
you have a (rare) ECC-NOR flash which may have 8 or 16 byte minimal
I/O unit size.

To address that, we're using biggest(4096 as of now) value to align actual
data chunk that has to be written to the flash. It will work for NOR as
well, and by the time we'll have to change that hard-coded value, it will
be exposed somewhere via /proc or /sys layers. To construct data chunk, tmpfs
is used, which should work even without writable partition available.

Signed-off-by: Vitaly Bordug <vitb@kernel.crashing.org>