Add Code {} directive to "tc language"
The current possibility to formulate test cases outside of the
expect-style "<command> <response>" style of specifying an external
script is too clumsy for short code pieces.

This patch adds a "Code sections" makeing it possible to embed pure
tcl in testcases, and have them run within testsuite context.  Code
sections should not use tcl commands to break commands flow (as
return, or exit).

Until we have proper "virtual" context_command functions, to access
the "target", the specific context_command fuctions need to be called.

The return value of the Code section is the content of the "res"
variable after the last statement, so be sure to set it.

Commands{} and Code{} blocks can be sequenced in a duts_tc and will be
executed in that order.

Signed-off-by: Vitaly Bordug <vitb@kernel.crashing.org>
Signed-off-by: Detlev Zundel <dzu@denx.de>