Use configuration variable for RAM address.
All the u-boot memory testcases in DUTS assume RAM starts at 0, which
is not the case for all boards (especially ARM boards). To be prepared
for this, switch all memory-related tests to use $CFG_RAM_WS_BASE.

As we now also do arithmetic calculations in Tcl with the DUTS
variables, it has become clear that we should use the "0x" prefix to
be consistent with Tcls default base.  This has been changed in all
the places touched by this commit but needs a separate follow-up patch
to convert all the rest.

Other small fixes, like ability to have different target name for
u-boot and kernel build, etc.

Signed-off-by: Detlev Zundel <dzu@denx.de>
Signed-off-by: Vitaly Bordug <vitb@kernel.crashing.org>