Get rid of Make<something> params in device descriptions
There is no reason to have parameters for the build process in a
separate namespace.  So we move everything we need to Vars adding a
BUILD_ prefix.

Make<something> routines were thus removed, together with explicit
check for this stuff in valid_devices().  The latter seemed bogus
anyway.

All the supported device descriptions were converted to the new
format, tests were run for some 'powerpc' devices.

Signed-off-by: Vitaly Bordug <vitb@kernel.crashing.org>
Signed-off-by: Detlev Zundel <dzu@denx.de>