Get 'host' context in line with other contexts.
We want to use the 'host' context to script the local host just as
we script the target board.  This is for example essential to setup
endpoints used from the device test.

For this we want one continous shell to execute our host commands.
So do not open one shell per single command, but per command block.
This probably needs some more changes to clean up correctly.

Signed-off-by: Detlev Zundel <dzu@denx.de>